requirements.txt
Django==2.0
Pillow
psycopg2
django-ckeditor